Added algorithms
