+++ server/controllers/mock.controller.ts
@@ -15,7 +15,7 @@ const log = useLog({
 export class MockController {
   constructor(
     private readonly mockService: MockService,
-    private readonly mockResponseService: MockResponseService,
+    private readonly mockResponseService: MockResponseService
   ) {}
 
   getRouter(): IRouterConfig {
@@ -41,7 +41,7 @@ export class MockController {
           const limit = +(req.params.limit ?? 1000);
           const requests = await this.mockService.getByUser(userId, limit);
           res.status(200).send(requests);
-        },
+        }
       )
       .post(
         /**
@@ -139,6 +139,7 @@ export class MockController {
             name,
             method,
             mockResponses,
+            selectedResponseId,
           } = req.body;
           const mock = await this.mockService.create(
             userId,
@@ -150,10 +151,11 @@ export class MockController {
             name,
             snifferId,
             mockResponses,
+            selectedResponseId
           );
 
           res.status(200).send(mock);
-        },
+        }
       );
 
     router
@@ -189,7 +191,7 @@ export class MockController {
 
           const mock = await this.mockService.getById(userId, mockId);
           res.status(200).send(mock);
-        },
+        }
       )
       .delete(
         /**
@@ -217,7 +219,7 @@ export class MockController {
           const { mockId } = req.params;
           const response = await this.mockService.delete(userId, mockId);
           res.status(200).send(response);
-        },
+        }
       )
       .patch(
         /**
@@ -299,11 +301,11 @@ export class MockController {
             headers,
             status,
             name,
-            snifferId,
+            snifferId
           );
 
           res.json(updatedMock).status(200);
-        },
+        }
       );
 
     router.route("/:mockId/selected-response").patch(
@@ -348,11 +350,11 @@ export class MockController {
         const mock = await this.mockService.setSelectedResponse(
           userId,
           mockId,
-          responseId,
+          responseId
         );
 
         res.status(200).send(mock);
-      },
+      }
     );
 
     router.route("/:mockId/activate").post(
@@ -383,10 +385,10 @@ export class MockController {
         const requests = await this.mockService.setIsActive(
           userId,
           mockId,
-          true,
+          true
         );
         res.status(200).send(requests);
-      },
+      }
     );
 
     router.route("/:mockId/deactivate").post(
@@ -418,10 +420,10 @@ export class MockController {
         const requests = await this.mockService.setIsActive(
           userId,
           mockId,
-          false,
+          false
         );
         res.status(200).send(requests);
-      },
+      }
     );
 
     return { router, path: "/sharkio/mocks" };
